requirements.txt
# Core LLM & Orchestration
langchain==0.3.25
langchain-openai==0.3.16
langchain-community==0.3.24
langgraph==0.4.1
openai==1.78.0

# Vector Stores
faiss-cpu==1.11.0
sentence-transformers==4.1.0
pinecone==6.0.0

# API Backend
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.4
python-multipart==0.0.20

# UI
streamlit==1.45.1
streamlit-chat==0.1.1

# Evaluation
ragas==0.2.15
datasets==3.6.0

# Data & Utilities
pypdf==5.4.0
python-docx==1.1.2
tiktoken==0.9.0
numpy==2.2.5
pandas==2.2.3
python-dotenv==1.1.0
httpx==0.28.1

# Monitoring & Logging
loguru==0.7.3
prometheus-client==0.22.0
pytest-asyncio==0.26.0
rank-bm25==0.2.2
langchain-nvidia-ai-endpoints==0.3.9
